guardian: move approval request formatting out of core
