+++ codex-rs/core/src/guardian/prompt.rs
@@ -11,9 +11,7 @@ use crate::compact::content_items_to_text;
 use crate::event_mapping::is_contextual_user_message_content;
 use crate::session::session::Session;
 use crate::session::turn_context::TurnContext;
-use codex_utils_output_truncation::approx_bytes_for_tokens;
 use codex_utils_output_truncation::approx_token_count;
-use codex_utils_output_truncation::approx_tokens_from_byte_count;
 
 use super::AUTO_REVIEW_DENIED_ACTION_APPROVAL_DEVELOPER_PREFIX;
 use super::GUARDIAN_MAX_MESSAGE_ENTRY_TOKENS;
@@ -23,8 +21,8 @@ use super::GUARDIAN_MAX_TOOL_TRANSCRIPT_TOKENS;
 use super::GUARDIAN_RECENT_ENTRY_LIMIT;
 use super::GuardianApprovalRequest;
 use super::GuardianAssessment;
-use super::TRUNCATION_TAG;
 use super::approval_request::format_guardian_action_pretty;
+pub(crate) use codex_guardian::guardian_truncate_text;
 
 /// Transcript entry retained for guardian review after filtering.
 #[derive(Debug, PartialEq, Eq)]
@@ -511,68 +509,6 @@ pub(crate) fn collect_guardian_transcript_entries(
     }
 
     entries
-}
-
-pub(crate) fn guardian_truncate_text(content: &str, token_cap: usize) -> (String, bool) {
-    if content.is_empty() {
-        return (String::new(), false);
-    }
-
-    let max_bytes = approx_bytes_for_tokens(token_cap);
-    if content.len() <= max_bytes {
-        return (content.to_string(), false);
-    }
-
-    let omitted_tokens = approx_tokens_from_byte_count(content.len().saturating_sub(max_bytes));
-    let marker = format!("<{TRUNCATION_TAG} omitted_approx_tokens=\"{omitted_tokens}\" />");
-    if max_bytes <= marker.len() {
-        return (marker, true);
-    }
-
-    let available_bytes = max_bytes.saturating_sub(marker.len());
-    let prefix_budget = available_bytes / 2;
-    let suffix_budget = available_bytes.saturating_sub(prefix_budget);
-    let (prefix, suffix) = split_guardian_truncation_bounds(content, prefix_budget, suffix_budget);
-
-    (format!("{prefix}{marker}{suffix}"), true)
-}
-
-fn split_guardian_truncation_bounds(
-    content: &str,
-    prefix_bytes: usize,
-    suffix_bytes: usize,
-) -> (&str, &str) {
-    if content.is_empty() {
-        return ("", "");
-    }
-
-    let len = content.len();
-    let suffix_start_target = len.saturating_sub(suffix_bytes);
-    let mut prefix_end = 0usize;
-    let mut suffix_start = len;
-    let mut suffix_started = false;
-
-    for (index, ch) in content.char_indices() {
-        let char_end = index + ch.len_utf8();
-        if char_end <= prefix_bytes {
-            prefix_end = char_end;
-            continue;
-        }
-
-        if index >= suffix_start_target {
-            if !suffix_started {
-                suffix_start = index;
-                suffix_started = true;
-            }
-            continue;
-        }
-    }
-
-    if suffix_start < prefix_end {
-        suffix_start = prefix_end;
-    }
-
-    (&content[..prefix_end], &content[suffix_start..])
 }
 
 /// The model is asked for strict JSON, but we still accept a surrounding prose
